LayoutTests/fast/css/background-layers-initial-size.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
div {
    background: linear-gradient(45deg, rgba(0,0,0,.4) 50%, transparent 0) 100% 0 / 25px 25px no-repeat,
                linear-gradient(-135deg, transparent 18px, yellowgreen 0);

    width: 12em;
    height: 8em;
}
</style>
</head>
<body>
<div id="testDiv"></div>
<script>
description("Test that 'initial' background-size values are properly handled in background layers.");

var computedBackgroundSize = window.getComputedStyle(testDiv)["background-size"];
// The second layer should have "auto" as background size as it is its initial value.
shouldBeEqualToString("computedBackgroundSize", "25px 25px, auto");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
