Safari-609.1.11.1/LayoutTests/fast/css/background-layers-initial-size-expected.txt

Test that 'initial' background-size values are properly handled in background layers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedBackgroundSize is "25px 25px, auto"
PASS successfullyParsed is true

TEST COMPLETE
